requirements.txt
openai==2.24.0
fastapi==0.133.1
streamlit==1.54.0
uvicorn==0.41.0
dotenv
pyyaml
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1